feat(Context Menu): Make global Context Menu functional
